requirements.txt
flask
pillow
pytubefix
pathlib
matplotlib==3.8.4
demucs
torch
torchaudio